backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
pydantic[email]
passlib[argon2]
python-jose[cryptography]